Move E2E for claim specific features into their own legacy test file
